requirements.txt
# Core dependencies
PyQt6>=6.6.0

# TOML support
# Python 3.10: tomli for reading
# Python 3.11+: tomllib is built-in
tomli>=2.0.0; python_version < "3.11"
# tomli_w is needed for writing TOML (all versions)
tomli-w>=1.0.0

# ELF file parsing (A2L address update)
pyelftools>=0.31

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0

# Packaging (optional, for building exe)
pyinstaller>=6.0.0
